[Refactoring] Remove RPCExecutor code duplication between the 2 consoles
