+++ src/qt/rpcconsole.h
@@ -36,18 +36,7 @@ public:
     explicit RPCConsole(QWidget* parent);
     ~RPCConsole();
 
-    static bool RPCExecuteCommandLine(std::string &strResult, const std::string &strCommand);
-
     void setClientModel(ClientModel* model);
-
-
-    enum MessageClass {
-        MC_ERROR,
-        MC_DEBUG,
-        CMD_REQUEST,
-        CMD_REPLY,
-        CMD_ERROR
-    };
 
 protected:
     virtual bool eventFilter(QObject* obj, QEvent* event);
